feat : advert Delete 기능 수정

+++ stad-spring/src/main/java/com/klpc/stadspring/domain/advert/service/AdvertService.java
@@ -172,10 +172,12 @@ public class AdvertService {
     @Transactional(readOnly = false)
     public DeleteAdvertResponse deleteAdvert(Long id){
         Advert advert = advertRepository.findById(id).orElseThrow(() -> new CustomException(ErrorCode.ENTITIY_NOT_FOUND));
-        Product product = advert.getProducts().get(0);
-        product.deleteProduct(product);
-        for(ProductType productType : product.getProductType()){
-            productType.deleteProductType(productType);
+        if(!advert.getProducts().isEmpty()) {
+            Product product = advert.getProducts().get(0);
+            product.deleteProduct(product);
+            for (ProductType productType : product.getProductType()) {
+                productType.deleteProductType(productType);
+            }
         }
         advert.deleteAdvert();
